requirements.txt
fastapi
uvicorn
psycopg2-binary
resend
pydantic[email]
python-dotenv
